fix(safety): critical security fixes, expanded tests, and doc cleanup
- Fix pyro channel race condition: single-fire with static flags prevents continuous HIGH
- Fix buffer overflows: strcpy→strncpy, sprintf→snprintf, serial input bounds check
- Fix integer overflow in GPS altitude conversion with float cast
- Fix EEPROM state validation with enum range check
- Fix memory leak in sensor factory with static allocation
- Implement watchdog reset detection via i.MX RT1062 SRC_SRSR register
- Add 25 unit tests: landing detection, altitude calculations, GPS validation, sensor health
- Extend test CSV from 48 to 62 fields to match data pipeline
- Update version references from v0.51 to v0.10.0 across docs
- Archive obsolete planning docs, remove duplicate COMPETITIVE_ANALYSIS.md

+++ src/ms5611_functions.cpp
@@ -153,7 +153,7 @@ bool ms5611_calibrate_with_gps(uint32_t timeout_ms) {
             float sea_level_Pa = STANDARD_SEA_LEVEL_PRESSURE * HPA_TO_PA_CONVERSION_FACTOR; // Convert from hPa to Pa
             
             float raw_altitude = HYPSOMETRIC_CONSTANT_A * (1.0 - pow(current_pressure_Pa / sea_level_Pa, HYPSOMETRIC_EXPONENT));
-            baro_altitude_offset = (GPS_altitude / MM_TO_M_CONVERSION_FACTOR) - raw_altitude;  // Convert GPS altitude from mm to m
+            baro_altitude_offset = (static_cast<float>(GPS_altitude) / MM_TO_M_CONVERSION_FACTOR) - raw_altitude;  // Convert GPS altitude from mm to m
             
             Serial.print(F("Calibration attempt #"));
             Serial.print(attempts);
